:muscle: Define `isOptionalOf/isStrictOf` in `factory.ts`
